requirements.txt
alembic==1.3.1
boto3==1.10.18
celery==4.3.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
flower==0.9.3
gunicorn==20.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
redis==3.3.11
requests==2.22.0
SQLAlchemy==1.3.11
SQLAlchemy-Utils==0.35.0